requirements.txt
asgiref==3.4.1
Django==3.2.9
django-braces==1.15.0
django-filter==21.1
Pillow==8.4.0
psycopg2==2.9.2
pyodbc==4.0.32
pytz==2021.3
sqlparse==0.4.2
Pillow==8.4.0
django-filter==21.1
django-ckeditor==6.2.0
whitenoise==5.3.0
gunicorn==20.1.0
dj-database-url==0.5.0
python-decouple==3.5
django-bootstrap-modal-forms==2.2.0 